requirements.txt
# Core
torch
torchvision
numpy
opencv-python
Pillow
albumentations
scikit-image
scikit-learn
matplotlib


# IO / formats
tifffile
imageio
imagecodecs


# CLI / config
pyyaml
rich
tqdm
scipy